+++ src/Microsoft.Extensions.DependencyInjection.Abstractions/Extensions/ServiceCollectionDescriptorExtensions.cs
@@ -2,310 +2,12 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using Microsoft.Extensions.DependencyInjection.Abstractions;
 
 namespace Microsoft.Extensions.DependencyInjection.Extensions
 {
-    public static class ServiceCollectionOrderedExtensions
-    {
-        public static IServiceCollection AddOrdered<TService, TImplementation>(this IServiceCollection services)
-          where TService : class
-          where TImplementation : class, TService
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            return AddOrdered(services, ServiceDescriptor.Transient(typeof(TService), typeof(TImplementation)));
-        }
-        public static IServiceCollection AddOrdered<TService>(
-            this IServiceCollection services,
-            TService implementationInstance)
-            where TService : class
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            if (implementationInstance == null)
-            {
-                throw new ArgumentNullException(nameof(implementationInstance));
-            }
-
-            return AddOrdered(services, ServiceDescriptor.Singleton(typeof(TService), implementationInstance));
-        }
-
-        public static IServiceCollection AddOrdered(
-            this IServiceCollection services,
-            Type serviceType,
-            object implementationInstance)
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            if (serviceType == null)
-            {
-                throw new ArgumentNullException(nameof(serviceType));
-            }
-
-            if (implementationInstance == null)
-            {
-                throw new ArgumentNullException(nameof(implementationInstance));
-            }
-
-            return AddOrdered(services, ServiceDescriptor.Singleton(serviceType, implementationInstance));
-        }
-
-        public static IServiceCollection AddOrdered(
-            this IServiceCollection services,
-            Type serviceType,
-            Type implementationType)
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            if (serviceType == null)
-            {
-                throw new ArgumentNullException(nameof(serviceType));
-            }
-
-            if (implementationType == null)
-            {
-                throw new ArgumentNullException(nameof(implementationType));
-            }
-
-            return AddOrdered(services, ServiceDescriptor.Transient(serviceType, implementationType));
-        }
-
-        public static IServiceCollection AddOrdered(
-           this IServiceCollection collection,
-           ServiceDescriptor descriptor)
-        {
-            if (collection == null)
-            {
-                throw new ArgumentNullException(nameof(collection));
-            }
-
-            if (descriptor == null)
-            {
-                throw new ArgumentNullException(nameof(descriptor));
-            }
-
-            var collectionDescriptor = GetOrderedDescriptor(collection, descriptor.ServiceType);
-            collectionDescriptor.Descriptors.Add(descriptor);
-            return collection;
-        }
-
-        private static OrderedServiceDescriptor GetOrderedDescriptor(
-            this IServiceCollection collection,
-            Type serviceType)
-        {
-            var descriptor = (OrderedServiceDescriptor)
-                collection.FirstOrDefault(d =>
-                    d.GetType() == typeof(OrderedServiceDescriptor) &&
-                    d.ServiceType == serviceType);
-            if (descriptor == null)
-            {
-                descriptor = new OrderedServiceDescriptor(serviceType);
-                collection.Add(descriptor);
-                var holderType = typeof(OrderedServiceDescriptorHolder<>).MakeGenericType(serviceType);
-                collection.AddSingleton(holderType,
-                    Activator.CreateInstance(holderType, descriptor));
-                collection.TryAddTransient(
-                    typeof(IOrdered<>).MakeGenericType(serviceType),
-                    typeof(Ordered<>).MakeGenericType(serviceType));
-            }
-            return descriptor;
-        }
-    }
-
-    class OrderedServiceDescriptorHolder<T>
-    {
-        public OrderedServiceDescriptorHolder(OrderedServiceDescriptor serviceDescriptor)
-        {
-            ServiceDescriptor = serviceDescriptor;
-        }
-
-        public OrderedServiceDescriptor ServiceDescriptor { get; }
-    }
-
-    internal class Ordered<T>: IOrdered<T>, IDisposable
-    {
-        private readonly IServiceProvider _serviceProvider;
-        private readonly OrderedServiceDescriptorHolder<T> _descriptorHolder;
-        private readonly Lazy<List<T>> _values;
-
-        public Ordered(IServiceProvider serviceProvider, OrderedServiceDescriptorHolder<T> descriptorHolder)
-        {
-            _serviceProvider = serviceProvider;
-            _descriptorHolder = descriptorHolder;
-            _values = new Lazy<List<T>>(() => GenerateValues().ToList());
-        }
-
-        private  IEnumerable<T> GenerateValues()
-        {
-            foreach (var descriptor in _descriptorHolder.ServiceDescriptor.Descriptors)
-            {
-                if (descriptor.ImplementationFactory != null)
-                {
-                    yield return (T) descriptor.ImplementationFactory(_serviceProvider);
-                }
-                else if (descriptor.ImplementationType != null)
-                {
-                    yield return (T) ActivatorUtilities.CreateInstance(_serviceProvider, descriptor.ImplementationType);
-                }
-                else
-                {
-                    yield return (T) descriptor.ImplementationInstance;
-                }
-            }
-        }
-
-        public IEnumerator<T> GetEnumerator()
-        {
-            return _values.Value.GetEnumerator();
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
-
-        public void Dispose()
-        {
-            if (_values.IsValueCreated)
-            {
-                foreach (var value in _values.Value)
-                {
-                    (value as IDisposable)?.Dispose();
-                }
-            }
-        }
-    }
-
-    public static class ServiceCollectionEnumerableExtensions
-    {
-        public static IServiceCollection AddEnumerable<TService, TImplementation>(this IServiceCollection services)
-          where TService : class
-          where TImplementation : class, TService
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            return AddEnumerable(services, ServiceDescriptor.Transient(typeof(TService), typeof(TImplementation)));
-        }
-        public static IServiceCollection AddEnumerable<TService>(
-            this IServiceCollection services,
-            TService implementationInstance)
-            where TService : class
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            if (implementationInstance == null)
-            {
-                throw new ArgumentNullException(nameof(implementationInstance));
-            }
-
-            return AddEnumerable(services, ServiceDescriptor.Singleton(typeof(TService), implementationInstance));
-        }
-
-        public static IServiceCollection AddEnumerable(
-            this IServiceCollection services,
-            Type serviceType,
-            object implementationInstance)
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            if (serviceType == null)
-            {
-                throw new ArgumentNullException(nameof(serviceType));
-            }
-
-            if (implementationInstance == null)
-            {
-                throw new ArgumentNullException(nameof(implementationInstance));
-            }
-
-            return AddEnumerable(services, ServiceDescriptor.Singleton(serviceType, implementationInstance));
-        }
-
-        public static IServiceCollection AddEnumerable(
-            this IServiceCollection services,
-            Type serviceType,
-            Type implementationType)
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            if (serviceType == null)
-            {
-                throw new ArgumentNullException(nameof(serviceType));
-            }
-
-            if (implementationType == null)
-            {
-                throw new ArgumentNullException(nameof(implementationType));
-            }
-
-            return AddEnumerable(services, ServiceDescriptor.Transient(serviceType, implementationType));
-        }
-
-        public static IServiceCollection AddEnumerable(
-           this IServiceCollection collection,
-           ServiceDescriptor descriptor)
-        {
-            if (collection == null)
-            {
-                throw new ArgumentNullException(nameof(collection));
-            }
-
-            if (descriptor == null)
-            {
-                throw new ArgumentNullException(nameof(descriptor));
-            }
-
-            var collectionDescriptor = GetEnumerableDescriptor(collection, descriptor.ServiceType);
-            collectionDescriptor.Descriptors.Add(descriptor);
-            return collection;
-        }
-
-        private static EnumerableServiceDescriptor GetEnumerableDescriptor(
-            this IServiceCollection collection,
-            Type serviceType)
-        {
-            var descriptor = (EnumerableServiceDescriptor)
-                collection.FirstOrDefault(d =>
-                    d.GetType() == typeof(EnumerableServiceDescriptor) &&
-                    d.ServiceType == serviceType);
-            if (descriptor == null)
-            {
-                descriptor = new EnumerableServiceDescriptor(serviceType);
-                collection.Add(descriptor);
-            }
-            return descriptor;
-        }
-    }
-
     public static class ServiceCollectionDescriptorExtensions
     {
         /// <summary>
