+++ src/Microsoft.Extensions.DependencyInjection.Abstractions/Extensions/ServiceCollectionEnumerableExtensions.cs
@@ -1,0 +1,145 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Linq;
+
+namespace Microsoft.Extensions.DependencyInjection.Extensions
+{
+    public static class ServiceCollectionEnumerableExtensions
+    {
+        public static IServiceCollection AddEnumerable<TService>(this IServiceCollection services)
+            where TService : class
+        {
+            if (services == null)
+            {
+                throw new ArgumentNullException(nameof(services));
+            }
+            AddEnumerable(services, typeof(TService));
+            return services;
+        }
+
+        public static IServiceCollection AddEnumerable(
+            this IServiceCollection services,
+            Type serviceType)
+        {
+            if (services == null)
+            {
+                throw new ArgumentNullException(nameof(services));
+            }
+            GetEnumerableDescriptor(services, serviceType);
+            return services;
+        }
+
+        public static IServiceCollection AddEnumerable<TService, TImplementation>(this IServiceCollection services)
+            where TService : class
+            where TImplementation : class, TService
+        {
+            if (services == null)
+            {
+                throw new ArgumentNullException(nameof(services));
+            }
+
+            return AddEnumerable(services, ServiceDescriptor.Transient(typeof(TService), typeof(TImplementation)));
+        }
+
+        public static IServiceCollection AddEnumerable<TService>(
+            this IServiceCollection services,
+            TService implementationInstance)
+            where TService : class
+        {
+            if (services == null)
+            {
+                throw new ArgumentNullException(nameof(services));
+            }
+
+            if (implementationInstance == null)
+            {
+                throw new ArgumentNullException(nameof(implementationInstance));
+            }
+
+            return AddEnumerable(services, ServiceDescriptor.Singleton(typeof(TService), implementationInstance));
+        }
+
+        public static IServiceCollection AddEnumerable(
+            this IServiceCollection services,
+            Type serviceType,
+            object implementationInstance)
+        {
+            if (services == null)
+            {
+                throw new ArgumentNullException(nameof(services));
+            }
+
+            if (serviceType == null)
+            {
+                throw new ArgumentNullException(nameof(serviceType));
+            }
+
+            if (implementationInstance == null)
+            {
+                throw new ArgumentNullException(nameof(implementationInstance));
+            }
+
+            return AddEnumerable(services, ServiceDescriptor.Singleton(serviceType, implementationInstance));
+        }
+
+        public static IServiceCollection AddEnumerable(
+            this IServiceCollection services,
+            Type serviceType,
+            Type implementationType)
+        {
+            if (services == null)
+            {
+                throw new ArgumentNullException(nameof(services));
+            }
+
+            if (serviceType == null)
+            {
+                throw new ArgumentNullException(nameof(serviceType));
+            }
+
+            if (implementationType == null)
+            {
+                throw new ArgumentNullException(nameof(implementationType));
+            }
+
+            return AddEnumerable(services, ServiceDescriptor.Transient(serviceType, implementationType));
+        }
+
+        public static IServiceCollection AddEnumerable(
+            this IServiceCollection collection,
+            ServiceDescriptor descriptor)
+        {
+            if (collection == null)
+            {
+                throw new ArgumentNullException(nameof(collection));
+            }
+
+            if (descriptor == null)
+            {
+                throw new ArgumentNullException(nameof(descriptor));
+            }
+
+            var collectionDescriptor = GetEnumerableDescriptor(collection, descriptor.ServiceType);
+            collectionDescriptor.Descriptors.Add(descriptor);
+            return collection;
+        }
+
+        private static EnumerableServiceDescriptor GetEnumerableDescriptor(
+            this IServiceCollection collection,
+            Type serviceType)
+        {
+            var descriptor = (EnumerableServiceDescriptor)
+                collection.FirstOrDefault(d =>
+                    d.GetType() == typeof(EnumerableServiceDescriptor) &&
+                    d.ServiceType == serviceType);
+            if (descriptor == null)
+            {
+                descriptor = new EnumerableServiceDescriptor(serviceType);
+                collection.Add(descriptor);
+            }
+            return descriptor;
+        }
+    }
+}
