+++ src/Microsoft.Extensions.DependencyInjection.Abstractions/OrderedServiceDescriptor.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Microsoft.Extensions.DependencyInjection
+{
+    public class OrderedServiceDescriptor: EnumerableServiceDescriptor
+    {
+        public OrderedServiceDescriptor(Type serviceType) : base(serviceType)
+        {
+        }
+    }
+}
